[clean] Removed checkings for very old Django versions: Django <= 1.8
