+++ django_dynamic_fixture/fixture_algorithms/random_fixture.py
@@ -15,8 +15,6 @@ except ImportError:
 
 try:
     from django.contrib.gis.geos import *
-except ImportError:
-    pass # Django < 1.7
 except ImproperlyConfigured:
     pass  # enviroment without geo libs
 except Exception:
