+++ django_dynamic_fixture/fixture_algorithms/tests/test_default_fixture.py
@@ -8,21 +8,15 @@ from django.conf import settings
 from django.core.exceptions import ImproperlyConfigured
 
 
-try:
-    from django.contrib.gis.geos import *
-except ImportError:
-    pass  # Django < 1.7
+from django.contrib.gis.geos import *
 try:
     from django.contrib.gis.db import models as geomodels
-except ImportError:
-    pass  # Django < 1.7
 except ImproperlyConfigured:
     pass  # enviroment without geo libs
 
 
 from django.test import TestCase
 
-from django_dynamic_fixture.django_helper import django_greater_than
 from django_dynamic_fixture.fixture_algorithms.default_fixture import BaseDataFixture
 
 
@@ -31,11 +25,10 @@ class BaseDataFixtureTestCase(TestCase):
         self.fixture = BaseDataFixture()
 
     def test_uuid(self):
-        if django_greater_than('1.8'):
-            assert isinstance(self.fixture.generate_data(models.UUIDField()), uuid.UUID)
+        assert isinstance(self.fixture.generate_data(models.UUIDField()), uuid.UUID)
 
 
-if django_greater_than('1.7') and (hasattr(settings, 'DDF_TEST_GEODJANGO') and settings.DDF_TEST_GEODJANGO):
+if (hasattr(settings, 'DDF_TEST_GEODJANGO') and settings.DDF_TEST_GEODJANGO):
     from django_dynamic_fixture.fixture_algorithms.default_fixture import GeoDjangoFixtureMixin
 
     # Mixing for tests
@@ -69,62 +62,3 @@ if django_greater_than('1.7') and (hasattr(settings, 'DDF_TEST_GEODJANGO') and s
 
         def test_geometrycollectionfield_config(self):
             assert isinstance(self.fixture.generate_data(geomodels.GeometryCollectionField()), GeometryCollection)
-
-
-if django_greater_than('1.8'):
-    try:
-        import psycopg2
-        from django.contrib.postgres.fields import ArrayField
-        from django_dynamic_fixture.fixture_algorithms.default_fixture import PostgresFixtureMixin
-        from django_dynamic_fixture.fixture_algorithms.sequential_fixture import SequentialDataFixture, StaticSequentialDataFixture
-        from django_dynamic_fixture.fixture_algorithms.random_fixture import RandomDataFixture
-        from django_dynamic_fixture.fixture_algorithms.unique_random_fixture import UniqueRandomDataFixture
-
-        class PostgresDataFixtureTestMixin(object):
-            def test_arrayfield_integer_config(self):
-                data = self.fixture.generate_data(ArrayField(models.IntegerField()))
-                assert isinstance(data, list)
-                assert isinstance(data[0], int)
-
-            def test_arrayfield_char_config(self):
-                data = self.fixture.generate_data(ArrayField(models.CharField()))
-                assert isinstance(data, list)
-                assert isinstance(data[0], six.text_type)
-
-            def test_arrayfield_datetime_config(self):
-                data = self.fixture.generate_data(ArrayField(models.DateTimeField()))
-                assert isinstance(data, list)
-                assert isinstance(data[0], datetime)
-
-            def test_arrayfield_email_config(self):
-                data = self.fixture.generate_data(ArrayField(models.EmailField(max_length=100)))
-                assert isinstance(data, list)
-                assert isinstance(data[0], six.text_type)
-
-
-        class PostgresSequentialDataFixtureTestCase(TestCase, PostgresDataFixtureTestMixin):
-            def setUp(self):
-                class CustomFixture(SequentialDataFixture, PostgresFixtureMixin):
-                    pass
-                self.fixture = CustomFixture()
-
-        class PostgresStaticSequentialDataFixtureTestCase(TestCase, PostgresDataFixtureTestMixin):
-            def setUp(self):
-                class CustomFixture(StaticSequentialDataFixture, PostgresFixtureMixin):
-                    pass
-                self.fixture = CustomFixture()
-
-        class PostgresRandomDataFixtureTestCase(TestCase, PostgresDataFixtureTestMixin):
-            def setUp(self):
-                class CustomFixture(RandomDataFixture, PostgresFixtureMixin):
-                    pass
-                self.fixture = CustomFixture()
-
-        class PostgresUniqueRandomDataFixtureTestCase(TestCase, PostgresDataFixtureTestMixin):
-            def setUp(self):
-                class CustomFixture(UniqueRandomDataFixture, PostgresFixtureMixin):
-                    pass
-                self.fixture = CustomFixture()
-
-    except ImportError:
-        print('Skipping Postgres tests because psycopg2 has not been installed.')
